// frontend/src/app/shared/services/api/models/message-type-enum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Zidnahum Hudaa Project API
 * OpenAPI spec version: 1.0.0
 */

export type MessageTypeEnum = (typeof MessageTypeEnum)[keyof typeof MessageTypeEnum];

export const MessageTypeEnum = {
  MEMO: 1,
  TEST: 2,
  ALNAWAWIA: 3,
  ALSAALIHIN: 4,
  ALLAH_NAMES: 5,
  ELITE_TEST: 6,
  VIEWING: 7,
  EXTRA_HADEETH: 8,
} as const;